pyproject.toml
[tool.poetry]
name = "wework-automation"
version = "0.1.0"
description = ""
authors = ["Vladimir Kavalerov <vladimir@kavalerov.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.12.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
